Refactor issue hygiene merge script into focused modules.
Split parsing, GitHub transport, project sync, issue operations, and orchestration into dedicated files, while keeping the legacy script as a compatibility shim and pointing the merge workflow to the new entrypoint.

Made-with: Cursor
