+++ .github/scripts/issue-hygiene/github-client.mjs
@@ -1,0 +1,58 @@
+/* global fetch */
+
+const REST_API_BASE = process.env.GITHUB_API_URL || 'https://api.github.com';
+const GRAPHQL_API_URL = process.env.GITHUB_GRAPHQL_URL || 'https://api.github.com/graphql';
+
+export async function githubRequest(path, init = {}) {
+  const token = process.env.GITHUB_TOKEN;
+  if (!token) {
+    throw new Error('Missing GITHUB_TOKEN.');
+  }
+
+  const response = await fetch(`${REST_API_BASE}${path}`, {
+    ...init,
+    headers: {
+      Accept: 'application/vnd.github+json',
+      Authorization: `Bearer ${token}`,
+      'Content-Type': 'application/json',
+      'User-Agent': 'kynd-web-issue-hygiene',
+      ...(init.headers || {}),
+    },
+  });
+
+  if (!response.ok) {
+    const body = await response.text();
+    throw new Error(`GitHub REST ${response.status} for ${path}: ${body}`);
+  }
+
+  if (response.status === 204) {
+    return null;
+  }
+
+  return response.json();
+}
+
+export async function graphqlRequest(query, variables) {
+  const token = process.env.GITHUB_TOKEN;
+  if (!token) {
+    throw new Error('Missing GITHUB_TOKEN.');
+  }
+
+  const response = await fetch(GRAPHQL_API_URL, {
+    method: 'POST',
+    headers: {
+      Authorization: `Bearer ${token}`,
+      'Content-Type': 'application/json',
+      'User-Agent': 'kynd-web-issue-hygiene',
+    },
+    body: JSON.stringify({ query, variables }),
+  });
+
+  const payload = await response.json();
+
+  if (!response.ok || payload.errors?.length) {
+    throw new Error(`GitHub GraphQL request failed: ${JSON.stringify(payload.errors || payload)}`);
+  }
+
+  return payload.data;
+}
